\name{meteo.data}
\alias{meteo.data}
\docType{data}
\title{
Meteorological at Brest (France) for January month from 1973 to 2013}
\description{
The data sets contains daily temperatures (degrees), daily precipitations (mm), mean wind (m/s) and mean pressure. Some data are missing. }
\usage{data(meteo.data)}


\source{ http://eca.knmi.nl/dailydata/index.php}
%\references{}
\examples{
data(meteo.data)
}
\keyword{datasets}