\name{PibDetteDemoc}
\alias{PibDetteDemoc}
\docType{data}
\title{
Annual GDP and Debt data 1970-2010}
\description{
Annual GDP and Debt data 1970-2010}

\usage{data(PibDetteDemoc)}
\format{
  A data frame with 3198 observations on the following 5 variables.
  \describe{
    \item{\code{year}}{ year }
    \item{\code{PIB}}{GDP}
    \item{\code{Dette}}{debt}
    \item{\code{Democratie}}{democratie indice}
    \item{\code{country}}{country}
  }
}

%\references{}
\examples{
data(PibDetteDemoc)
## maybe str(PibDetteDemoc) 
}
\keyword{datasets}
